Rename UTs as tested files were renamed as well
