+++ impl/maven-cli/src/test/java/org/apache/maven/cling/invoker/mvn/local/LocalMavenInvokerTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.maven.cling.invoker.mvn.resident;
+package org.apache.maven.cling.invoker.mvn.local;
 
 import java.nio.file.FileSystem;
 import java.nio.file.Path;
@@ -36,13 +36,12 @@ import org.junit.jupiter.api.io.CleanupMode;
 import org.junit.jupiter.api.io.TempDir;
 
 /**
- * Resident UT.
+ * Local UT.
  */
-public class DefaultResidentMavenInvokerTest extends MavenInvokerTestSupport {
-
+public class LocalMavenInvokerTest extends MavenInvokerTestSupport {
     @Override
     protected Invoker createInvoker() {
-        return new ResidentMavenInvoker(ProtoLookup.builder()
+        return new LocalMavenInvoker(ProtoLookup.builder()
                 .addMapping(ClassWorld.class, new ClassWorld("plexus.core", ClassLoader.getSystemClassLoader()))
                 .build());
     }
